refactor: getUserResponse.ts를 user.ts로 이름 변경
